requirements/development.txt
-r base.txt

# factory_boy and faker to generate dummy data in the database
factory_boy
faker

# Flit to simplify package management
flit

# Hypothesis to find edge cases
hypothesis[pytz]

# ipdb to access the IPython debugger as if it were the pdb module
ipdb


# Mutatest to check if small modifications (mutations) in the code are detectable by tests
mutatest

# Mypy to support typing
mypy

# Pytest to run the tests
pytest

# pytest-cov to measure code coverage
pytest-cov

# Responses to mock requests
responses

# Tox to automate and standardize testing in Python
tox
